feat(rwd): MyPage 전역에서 반응형 state 관리
